wip refactor document_test.exs
