+++ .gitattributes
@@ -1,0 +1,5 @@
+*     text=auto
+
+*.ex  diff=elixir
+*.exs diff=elixir
+
