+++ config/config.exs
@@ -1,10 +1,9 @@
 import Config
 
 if config_env() in [:dev, :test] do
-  config :ex_typesense,
+  config :open_api_typesense,
     api_key: "xyz",
     host: "localhost",
     port: 8108,
-    scheme: "http",
-    options: %{}
+    scheme: "http"
 end
